Implement onboarding

+++ lib/views/login/verification.dart
@@ -1,7 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:lottie/lottie.dart';
 import 'package:piwo/config/theme/custom_colors.dart';
 import 'package:piwo/services/account.dart';
 import 'package:piwo/services/auth.dart';
+import 'package:piwo/services/onboarding.dart';
+import 'package:piwo/views/home/home_view.dart';
+import 'package:piwo/widgets/custom_scaffold.dart';
 import 'package:piwo/widgets/notifiers/login_notifier.dart';
 import 'package:provider/provider.dart';
 
@@ -22,25 +26,21 @@ class VerificationPage extends StatefulWidget {
 class VerificationPageState extends State<VerificationPage> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return CustomScaffold(
       body: Center(
         child: Padding(
-          padding: const EdgeInsets.only(top: 200.0, left: 20.0, right: 20.0),
+          padding: const EdgeInsets.only(top: 100.0, left: 20.0, right: 20.0),
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
-              Icon(
+              Lottie.asset(
                 !widget.isComfired
-                    ? Icons.help
+                    ? "assets/images/waiting.json"
                     : widget.isApproved
-                        ? Icons.check
-                        : Icons.cancel,
-                color: !widget.isComfired
-                    ? Colors.grey
-                    : widget.isApproved
-                        ? Colors.green
-                        : Colors.red,
-                size: 56,
+                        ? "assets/images/check.json"
+                        : "assets/images/denied.json",
+                width: 1000,
+                height: 275,
               ),
               const SizedBox(
                 height: 10,
@@ -56,28 +56,48 @@ class VerificationPageState extends State<VerificationPage> {
                 ),
               ),
               const SizedBox(height: 40.0),
-              MaterialButton(
-                minWidth: double.maxFinite,
-                color: CustomColors.themePrimary,
-                onPressed: () async {
-                  await AuthService().signOut();
+              if (!widget.isApproved) ...[
+                MaterialButton(
+                  minWidth: double.maxFinite,
+                  color: CustomColors.themePrimary,
+                  onPressed: () async {
+                    OnboardingService.saveOnboardingCompleted(false);
+                    await AuthService().signOut();
 
-                  if (!context.mounted) return;
-                  context.read<LoginStateNotifier>().logOut();
-                },
-                child: const Text('Uitloggen'),
-              ),
-              MaterialButton(
-                minWidth: double.maxFinite,
-                color: Colors.red,
-                onPressed: () async {
-                  AccountService().deleteAccount();
+                    if (!context.mounted) return;
+                    context.read<LoginStateNotifier>().logOut();
+                  },
+                  child: const Text('Uitloggen'),
+                ),
+                MaterialButton(
+                  minWidth: double.maxFinite,
+                  color: Colors.red,
+                  onPressed: () async {
+                    OnboardingService.saveOnboardingCompleted(false);
+                    AccountService().deleteAccount();
 
-                  if (!context.mounted) return;
-                  context.read<LoginStateNotifier>().logOut();
-                },
-                child: const Text('Verwijder je account'),
-              ),
+                    if (!context.mounted) return;
+                    context.read<LoginStateNotifier>().logOut();
+                  },
+                  child: const Text('Verwijder je account'),
+                ),
+              ] else ...[
+                MaterialButton(
+                  minWidth: double.maxFinite,
+                  color: CustomColors.themePrimary,
+                  onPressed: () async {
+                    if (!context.mounted) return;
+                    Navigator.pushAndRemoveUntil(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => const HomeView(),
+                      ),
+                      (route) => false,
+                    );
+                  },
+                  child: const Text('Ga naar de app'),
+                ),
+              ]
             ],
           ),
         ),
